+++ examples/PickingPainterMethod/PickingPainterMethod.cpp
@@ -58,16 +58,16 @@ void PickingPainterMethod::Init()
         
         // Create Materials
         UnselectedMaterial = new GenericShaderMaterial(ShaderUsage::Color | ShaderUsage::Diffuse);
-        UnselectedMaterial->SetColor(Vec4(0.8,0.8,0.8,1.0));
+        UnselectedMaterial->SetColor(Vec4(0.8f,0.8f,0.8f,1.f));
         
         SelectedMaterial = new GenericShaderMaterial(ShaderUsage::Color);
-        SelectedMaterial->SetColor(Vec4(1,1,0,1));
+        SelectedMaterial->SetColor(Vec4(1.f,1.f,0.f,1.f));
         
         // Set Selected Mesh PTR NULL
         SelectedMesh = NULL;
         
         // Create Cubes
-        srand( time( NULL ) );
+        srand( (unsigned int)time( NULL ) );
         
         // Create Geometry
         cubeHandle = new Cube(10,10,10);
@@ -88,7 +88,7 @@ void PickingPainterMethod::Init()
             // Add GameObject to Scene
             Scene->Add(Cube);
             // Set Random Position to the GameObject
-            Cube->SetPosition(Vec3((rand() % 200) -100,(rand() % 200)-100,(rand() % 200) -100));
+            Cube->SetPosition(Vec3((f32)(rand() % 200) -100.f, (f32)(rand() % 200)-100.f, (f32)(rand() % 200) -100.f));
         }
 
         // Create On Mouse Press Event
